requirements.txt
proxy-tools==0.1.0
pyobjc-core==9.0
pyobjc-framework-Cocoa==9.0
pyobjc-framework-WebKit==9.0
pywebview==3.7
